requirements.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters

faiss-cpu
pypdf

fastapi
uvicorn

streamlit
requests

python-dotenv
pandas

sentence-transformers
rank-bm25

ragas
pytest